+++ packages/tools/atlas/src/bin/base/world/run.sh
@@ -4,6 +4,6 @@ echo '
 ==== world ====
 ' &&
 
-node -r esm src/bin/base/world/01_world_110m_iso_a2_topojson.js &&
+node src/bin/base/world/01_world_110m_iso_a2_topojson.js &&
 
 echo